Fix Huawei vendor security patch display
Huawei Firmware B552 still shows same vendor patch level as B550,
because the respective prop ro.build.version.security_patch still
reflects the B550 version.

The respective prop with the correct information is:
 ro.huawei.build.version.security_patch

As a work-around, try to read the huawei prop and use its content,
if the prop exists and provides values.

Change-Id: I0fe2039e687c64445cc2c99bc7c409caaf8e7c78